fix: move conversion code to dedicatd package
