+++ pkg/db/migration/convert/convert000015/main.go
@@ -551,8 +551,4 @@ func (m *Migration) Migrate() error {
 }
 
 // Migration executes code along with the 15th database schema revision.
-// NOTE: for new migrations, when possible, it is best to define a <version>.go
-// in the `migration` package, next to the <version>_init.up.sql script. In
-// that case, the migration struct can be unexported and have a descriptive
-// type name.
 type Migration struct{}
